+++ ktlint-ruleset-experimental/src/test/kotlin/com/pinterest/ktlint/ruleset/experimental/ArgumentListWrappingRuleTest.kt
@@ -589,4 +589,99 @@ class ArgumentListWrappingRuleTest {
             )
         ).isEmpty()
     }
+
+    // https://github.com/pinterest/ktlint/issues/1081
+    @Test
+    fun testManyCorrections() {
+        assertThat(
+            ArgumentListWrappingRule().format(
+                """
+                package com.foo
+
+                class MyClass() {
+                    private fun doSomething() {
+                        if (d == 0 || e == 0f) {
+                            c.ee(hh, d, d, yy)
+                        } else {
+                            foo.blah()
+                            val dr = t - u
+                            val xx = -gg(dr) * rr(2f * d, dr.hh)
+                            foo.bar(
+                                dd,
+                                g - d
+                            )
+                            foo.baz(
+                                a.b, a.c - d
+                            )
+                            foo.biz(
+                                a.b, a.c - d,
+                                a.b, a.c,
+                                a.b + d, a.c
+                            )
+                            foo.baz(
+                                a.x - d, a.c
+                            )
+                            foo.biz(
+                                a.x - d, a.c,
+                                a.x, a.c,
+                                a.x, a.c - d
+                            )
+                            foo.baz(
+                                a.x, a.j + d
+                            )
+                        }
+                    }
+                }
+                """.trimIndent()
+            )
+        ).isEqualTo(
+            """
+            package com.foo
+
+            class MyClass() {
+                private fun doSomething() {
+                    if (d == 0 || e == 0f) {
+                        c.ee(hh, d, d, yy)
+                    } else {
+                        foo.blah()
+                        val dr = t - u
+                        val xx = -gg(dr) * rr(2f * d, dr.hh)
+                        foo.bar(
+                            dd,
+                            g - d
+                        )
+                        foo.baz(
+                            a.b,
+                            a.c - d
+                        )
+                        foo.biz(
+                            a.b,
+                            a.c - d,
+                            a.b,
+                            a.c,
+                            a.b + d,
+                            a.c
+                        )
+                        foo.baz(
+                            a.x - d,
+                            a.c
+                        )
+                        foo.biz(
+                            a.x - d,
+                            a.c,
+                            a.x,
+                            a.c,
+                            a.x,
+                            a.c - d
+                        )
+                        foo.baz(
+                            a.x,
+                            a.j + d
+                        )
+                    }
+                }
+            }
+            """.trimIndent()
+        )
+    }
 }